Fix SaitEH Run 1 by fixing XML config and adding return home for HMIP6
AR XML config fixed so HA bit adv.
R bit for router's global address. This needs to map to HA address.
Especially if static HA is set to that router the two need to be same
otherwise will not recognise returning home. Cannot have more than one as we only
associate one global address with mipv6 router.
Also Map address needs to be global and should not be same as HA.

Removed triggered dests when returning home and
no creation of reverse tunnels.