Added baseSettings ned parameter to RoutingTable6. Assertion removed in MStateMobileNode
baseSettings is needed because whenever we linkUpTrigger is true in ini file this
causes L2 Up/Down events to be recorded in cOutVectors. This requires that the
node have Mobile IPv6 enabled otherwise no callback function is not set
in WirelessEtherModule.

idea of baseSettings is to allow many different groups of nodes to share a similar
node settings instead of having to explicitly exactly same info for each in xml file.
Thus only one xml setting for a group mapped by the ini file. However this only
occurs for node settings though and would not work or make sense for routes.

MStateMobileNode includes changes that could cause instability. Still have not
found out why the assertion fails for CampusEH scenario.

Also the vectors recorded at MStateMobileNode were incomplete and in error
because the bule (binding update list entry) would be removed due to timeout or
when MN returned home anyway. To overcome this problem make sure it does not
return home otherwise results will be incomplete!!!!