Fixed buid issues with cmake 2.4 and omnetpp 3.2p1 which links in mpi by default
Trying to build this on supersupreme was a nightmare because cmake 2.4 made it
impossible to locate the correct libs automatically. It used the default path
before the custom path I detected. So we get 3.2p1 headers but libs in /usr/lib
from 3.0cvs rpm. This made me rebuild omnet thinking because gcc was updated
recently in June maybe undefined ref because ABI changed. Now omnet finds Lam
which was recently installed and so now omnet exe's require mpi libs. So
modified to cope with link in mpi as libsim_std contains what used to be
libsim_mpi I guess.

Will not work with omnet 2 now. May work with 3.1 not sure about mpi there since
I never have mpi installed on system.