resurrect MIPv6Network -r1 for stage 2 but mn fails to send NA no idea why
Changed DTD to allow empty interface elements: noticed that iface names do not
determine which iface is configured but purely based on the number of ifaces
in xml file. This was not behaviour I remembered. This means if we want to
configure eth2 we have to create an empty eth1.