Fix test_multiconfig_script.yaml wcmc y_22 r86 mn disconnects after 350
around the association to ap3 at 351.4 there is an ack for assoc req. that is
lost due to collision with ap2. This causes mn to resend assoc req. down
the line and subsequently mn fails to attach to any further APs. Problem
occurs due to noauth option on. But I think is also possible even for
noauth however I do not have an example for that. Read review.org under
todays date for further details.

This fix however adds approx. 500ms rescan and handover due to chain of
events. Not sure if real spec handles collisions to acks of controls
packets like so. Best fix for no handover is if mn did not resend assoc.
req. after assoc. resp received.